ai-backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pymupdf==1.25.4
python-docx==1.1.2
spacy==3.8.4
openai==1.66.3
requests==2.32.3
loguru==0.7.3
pydantic==2.10.6
typer==0.15.2
httpx==0.28.1
numpy==1.26.4
pymongo[srv]==4.6.1